+++ src/routes/snapshot/+page.svelte
@@ -23,7 +23,6 @@
     enhance,
     delayed,
     message,
-    empty,
     constraints,
     capture,
     restore